requirements.txt
pandas>=2.2.3,<3.0.0
PyYAML>=6.0.2,<7.0.0
requests>=2.32.3,<3.0.0
snowflake-connector-python>=3.12.2,<4.0.0
streamlit>=1.41.1,<2.0.0
pyspark>=3.5.3,<4.0.0
